move rawColors to separate file to fix problematic d.ts generation (default/colors as namespace)
